type mismatch

+++ loading_pipeline/lib/tasks/exports/write_new_entries_parquet.py
@@ -98,7 +98,7 @@ class WriteNewEntriesParquetTask(BaseWriteParquetTask):
         )
         if 'xpos' not in existing_annotations_ht.row:
             existing_annotations_ht = existing_annotations_ht.annotate(
-                xpos=hl.int64(xpos(existing_annotations_ht)),
+                xpos=hl.int32(xpos(existing_annotations_ht)),
             )
 
         annotations_ht = annotations_ht.union(
